MOD: cpu code is fixed

+++ db.h
@@ -9,34 +9,34 @@
 class Particle {
 	
 	private:
-		double q;	//charge of the particle
-		double m;	//mass of the particle
-		double x[2];	//array of the positions
-		double v[2];	//array of the velocities
-		double E[2];	//array of field
-		double I[2]; 	//array of interaction
+		float q;	//charge of the particle
+		float m;	//mass of the particle
+		float x[2];	//array of the positions
+		float v[2];	//array of the velocities
+		float E[2];	//array of field
+		float I[2]; 	//array of interaction
 
 		void set_field();
 
 	public:
 		Particle();
-		double get_mass();
-		double get_charge();
+		float get_mass();
+		float get_charge();
 
-		double* get_position();
-		double* get_velocity();
-		double* get_field();
+		float* get_position();
+		float* get_velocity();
+		float* get_field();
 		
 		void update_field(int N, int index, Particle * particles);
-		void set_position(double * position);
-		void set_velocity(double* velocity);
+		void set_position(float * position);
+		void set_velocity(float* velocity);
 		void set_interaction(int N, int index, Particle * particles);
-		void solve_time_step(double dt);
+		void solve_time_step(float dt);
 
 };
 
-void update_position(double dt, double T, const int N, Particle * particles);
-void electricField(double* E, double* x);
+void update_position(float dt, float T, const int N, Particle * particles);
+void electricField(float* E, float* x);
 void initial_condition(Particle * particles,int N);
 
 #endif
